# apps/cvs/cv_service/openapi.yaml
openapi: 3.0.0
info:
  title: CandidateV CV Management Service API
  version: 1.0.0
  description: API for creating, editing, and managing CVs
  contact:
    name: CandidateV Support
    email: support@candidatev.com
servers:
  - url: http://localhost:8002
    description: Local development server
  - url: https://{cv_service_url}
    description: Production server
    variables:
      cv_service_url:
        default: cv-service.candidatev.com
components:
  schemas:
...
    CVTemplate:
      type: object
      required:
        - id
        - name
        - category
      properties:
        id:
          type: string
          description: Template identifier
          example: professional
        name:
          type: string
          description: Template display name
          example: Professional
        preview_image_url:
          type: string
          nullable: true
          format: uri
          description: URL to template preview image
          example: https://example.com/templates/professional.jpg
        description:
          type: string
          nullable: true
          description: Template description
          example: A clean and professional CV template
        category:
          type: string
          description: Template category
          example: Business
        style_options:
          type: object
          description: Template styling options
          example:
            color_scheme: blue
            font_family: Roboto
            layout: standard
        is_premium:
          type: boolean
          default: false
          description: Whether this is a premium template
          example: false
...
    CV:
      type: object
      required:
        - id
        - user_id
        - metadata
        - content
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: CV unique identifier
          example: 123e4567-e89b-12d3-a456-426614174000
        user_id:
          type: string
          format: uuid
          description: Owner user ID
          example: 123e4567-e89b-12d3-a456-426614174001
        metadata:
          $ref: '#/components/schemas/CVMetadata'
        content:
          $ref: '#/components/schemas/CVContent'
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: 2025-04-10T12:00:00Z
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: 2025-04-10T15:30:00Z
    CVCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: CV title
          example: My Professional CV
        description:
          type: string
          nullable: true
          description: CV description
          example: CV for software engineering positions
        is_default:
          type: boolean
          default: false
          description: Whether this should be the user's default CV
          example: false
        template_id:
          type: string
          default: default
          description: Template identifier
          example: professional
        base_cv_id:
          type: string
          nullable: true
          format: uuid
          description: ID of an existing CV to copy content from
          example: null
    CVUpdateMetadata:
      type: object
      properties:
        name:
          type: string
          nullable: true
          description: CV title
          example: Updated CV Title
        description:
          type: string
          nullable: true
          description: CV description
          example: Updated description for my CV
        is_default:
          type: boolean
          nullable: true
          description: Whether this should be the user's default CV
          example: true
    CVUpdateContent:
      type: object
      properties:
        template_id:
          type: string
          nullable: true
          description: Template identifier
          example: modern
        style_options:
          type: object
          nullable: true
          description: CV style customization options
          example:
            color_scheme: teal
            font_family: Montserrat
        personal_info:
          type: object
          nullable: true
          description: Personal information section
          example:
            full_name: John Doe
            email: john.doe@example.com
            phone: +1 (123) 456-7890
        summary:
          type: string
          nullable: true
          description: CV summary or objective
          example: Experienced software engineer with 5+ years of experience in web development.
...
    CVListResponse:
      type: object
      required:
        - items
        - total
        - page
        - limit
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - id
              - title
              - last_modified
              - created_at
            properties:
              id:
                type: string
                format: uuid
                description: CV unique identifier
                example: 123e4567-e89b-12d3-a456-426614174000
              title:
                type: string
                description: CV title
                example: My Professional CV
              template_id:
                type: string
                description: Template identifier
                example: professional
              last_modified:
                type: string
                format: date-time
                description: Last modification timestamp
                example: 2025-04-10T15:30:00Z
              created_at:
                type: string
                format: date-time
                description: Creation timestamp
                example: 2025-04-10T12:00:00Z
        total:
          type: integer
          minimum: 0
          description: Total number of CVs
          example: 5
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        limit:
          type: integer
          minimum: 1
          description: Number of items per page
          example: 10
    TemplateListResponse:
      type: object
      required:
        - items
        - total
        - page
        - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CVTemplate'
        total:
          type: integer
          minimum: 0
          description: Total number of templates
          example: 3
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        limit:
          type: integer
          minimum: 1
          description: Number of items per page
          example: 10
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum: 
            - error
          description: Error status
          example: error
        message:
          type: string
          description: Error message
          example: CV not found
        code:
          type: string
          nullable: true
          description: Error code
          example: NOT_FOUND
        requestId:
          type: string
          nullable: true
          description: Request ID for tracing
          example: 123e4567-e89b-12d3-a456-426614174000
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
paths:
...
  /api/cv:
    get:
      summary: List user's CVs
      description: Retrieve all CVs for the authenticated user
      tags:
        - CV Management
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: List of CVs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CVListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      summary: Create a new CV
      description: Create a new CV, optionally based on an existing one
      tags:
        - CV Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CVCreate'
      responses:
        '201':
          description: CV created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CV'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cv/{cv_id}:
    parameters:
      - name: cv_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: CV unique identifier
    
    get:
      summary: Get CV
      description: Retrieve a specific CV
      tags:
        - CV Management
      security:
        - BearerAuth: []
      responses:
        '200':
          description: CV details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CV'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      summary: Delete CV
      description: Delete a specific CV
      tags:
        - CV Management
      security:
        - BearerAuth: []
      responses:
        '204':
          description: CV deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cv/{cv_id}/metadata:
    parameters:
      - name: cv_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: CV unique identifier
    
    put:
      summary: Update CV metadata
      description: Update CV title, description, and default status
      tags:
        - CV Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CVUpdateMetadata'
      responses:
        '200':
          description: CV updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CV'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cv/{cv_id}/content:
    parameters:
      - name: cv_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: CV unique identifier
    
    put:
      summary: Update CV content
      description: Update CV content, including template, personal info, and sections
      tags:
        - CV Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CVUpdateContent'
      responses:
        '200':
          description: CV content updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CV'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cv/{cv_id}/duplicate:
    parameters:
      - name: cv_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: CV unique identifier to duplicate
    
    post:
      summary: Duplicate CV
      description: Create a new CV based on an existing one
      tags:
        - CV Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CVCreate'
      responses:
        '201':
          description: CV duplicated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CV'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cv/templates:
    get:
      summary: List CV templates
      description: Retrieve all available CV templates
      tags:
        - Templates
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Number of items per page
        - name: category
          in: query
          schema:
            type: string
          description: Filter templates by category
      responses:
        '200':
          description: List of CV templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cv/templates/{template_id}:
    parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
        description: Template identifier
    
    get:
      summary: Get template
      description: Retrieve a specific CV template
      tags:
        - Templates
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CVTemplate'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' 